New Art Inspirations
This Commit starts the new pixel style.
This allows for a more flexible and humorous
visual representation of the gameplay. I feel this is a better
direction because it not only saves time and effort on player art,
it also allows for higher quality visuals when more time can be
spent on easilly modifiable and adjustable frame sets, using a
low poly but still visually clear style.
The style is inspired by paper mario thousand year door and the Game & Watch character in Super Smash Melee.