Time Freeze is done?
I got the everything "working" when the time freeze powerup is collected
the distortion now freezes time in a small radius around the player,
locking all movement. Attacks can still be spawned, and their momentum
is stored, you can create a pretty cool wave of attacks with this.
However, I think theres going to be a bug when another player used a
speed up distortion into a feeze distortion. I'm done for now, I'll work
with Brian on the multiplier in our meeting. This is too long of a
commit log